commented the rest of the code

+++ src/platforms/win64_platform.hpp
@@ -6,21 +6,25 @@
 #include <algorithm>
 
 #include "../common/simulate.hpp"
+#include "../common/types/types.hpp"
 
 int WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int nShowCmd);
 
 namespace Windows {
+    // This bool gets set to false whenever you press the x button
     bool going = true;
+    // Here all the pixels are stored
     COLORREF* frameBuffer;
-    long bufferWidth;
-    long bufferHeight;
+    i32 bufferWidth;
+    i32 bufferHeight;
+    // This contains some info about the buffer
     BITMAPINFO bmpInfo;
 
     void runFrame(HWND window, HDC handle);
     void processInput(HWND window);
     void renderFrame(HDC handle);
 
-    int WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int nShowCmd);
+    i32 WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, i32 nShowCmd);
     LRESULT windowCallback(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 }
